fix(api): separate Sentry DSNs for frontend/backend and fix item similarity 500 (#335)
Separate SENTRY_FRONTEND_DSN and SENTRY_BACKEND_DSN so each service
reports to its own Sentry project. The backend falls back to SENTRY_DSN
for backwards compatibility.

Fix item similarity 500 caused by reading PostgreSQL COUNT(*) bigint
columns with GetInt32 (which throws InvalidCastException). Changed to
GetInt64 with safe cast.

Closes RECEIPTS-467

+++ src/Infrastructure/Services/ReportService.cs
@@ -191,9 +191,9 @@ public partial class ReportService(IDbContextFactory<ApplicationDbContext> conte
 			{
 				edges.Add(new DescriptionSimilarityEdge(
 					reader.GetString(0),
-					reader.GetInt32(1),
+					(int)reader.GetInt64(1),
 					reader.GetString(2),
-					reader.GetInt32(3),
+					(int)reader.GetInt64(3),
 					reader.GetDouble(4)));
 			}
 		}
